Render blocks using converted content
