+++ pages/src/_plugins/filters/workers/worker_images.rb
@@ -20,7 +20,7 @@ module Jekyll
             end
 
             # Transform the images into the proper HTML string
-            return source_images.collect do |worker_image|
+            source_images.collect do |worker_image|
                 pieces = File.basename(worker_image, '.*').split('-')
                 ext = File.extname(worker_image)
