+++ pages/src/_plugins/templates/base_block.rb
@@ -1,6 +1,4 @@
-Arguments = Struct.new(:keyed, :unkeyed)
-
-class BaseTag < Liquid::Tag
+class BaseBlock < Liquid::Block
     def initialize(tag_name, args, tokens)
         super
         @arguments = args.strip.split(" ")
@@ -8,12 +6,19 @@ class BaseTag < Liquid::Tag
     end
 
     def render(context)
+        content = super
         arguments = parse_arguments(context)
-        return render_tag(context, arguments)
+        return render_block(context, content, arguments)
     end
 
-    def render_tag(context, arguments)
+    def render_block(context, arguments)
         return ""
+    end
+
+    def convert_content(context, content)
+        page = context.registers[:site].pages.detect { |p| p.path==context['page']['path'] }
+        renderer = Jekyll::Renderer.new(context.registers[:site], page)
+        return renderer.convert(content)
     end
 
     private
