+++ pages/src/_plugins/utils/require.rb
@@ -1,0 +1,2 @@
+require_relative "building_utils"
+require_relative "worker_utils"
